+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## master
+* Add support for lazy loaded page in `admin-panel` template
+
 ## 4.0.3
 * Fix login in `admin-panel` template